requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
pydantic>=2.0.0
optuna>=3.0.0
numpy>=1.20.0
pandas>=1.3.0
sqlalchemy>=2.0.0
plotly>=5.0.0
python-dotenv>=0.19.0
gunicorn>=20.0.0
